remove misleading error message
This was changed because of the following:
"""
WARNING  Resources.inventory_resource:inventory_resource.py:68
Integrity Error, this inventory is already in the database.
Error: (psycopg2.errors.ForeignKeyViolation) insert or update on table "inventory" violates foreign key constraint "inventory_trip_id_fkey"
DETAIL:  Key (trip_id)=(999) is not present in table "trip".
"""

The hardcoded error message "Integrity Error, this inventory is already in the database" is misleading, and hides (kind of) the real error.
If in hurry we might end up looking for the wrong error to fix.