git add login endpoint
This commit will add a login endpoint that verifies the user password, and generates a jwt token on a successfull login.
This jwt token will later be used to authenticate the user, and has a session limit of 45 minutes.

We need to generate and save a SECRET_KEY in the env.